feat: add redirect from `/` to default locale
