+++ app/[lang]/layout.tsx
@@ -1,26 +1,12 @@
 import { i18n, type Locale } from '@/i18n-config';
-import type { Metadata } from 'next';
-import { Nunito_Sans } from 'next/font/google';
 import Footer from './Footer';
-import './globals.css';
 import Navbar from './Navbar';
-
-const nunitoSans = Nunito_Sans({
-  subsets: ['latin'],
-  display: 'swap',
-});
-
-export const metadata: Metadata = {
-  title: 'Virufy: Developing Cutting Edge AI Technology in Healthcare',
-  description:
-    'Virufy is building an App for early detection symptoms analyzing vocal patterns in minutes. Our goal is a healthier world at your fingertips.',
-};
 
 export async function generateStaticParams() {
   return i18n.locales.map((locale) => ({ lang: locale }));
 }
 
-export default async function RootLayout({
+export default async function LangLayout({
   children,
   params: { lang },
 }: Readonly<{
@@ -28,12 +14,10 @@ export default async function RootLayout({
   params: { lang: Locale };
 }>) {
   return (
-    <html lang={lang}>
-      <body className={nunitoSans.className}>
-        <Navbar lang={lang} />
-        <main>{children}</main>
-        <Footer lang={lang} />
-      </body>
-    </html>
+    <>
+      <Navbar lang={lang} />
+      <section>{children}</section>
+      <Footer lang={lang} />
+    </>
   );
 }
